Update config and views
The way the configuration is handled has changed. However, the documentation has not been impact yet. This will come in the next commit. Also, the pretty-scaler has been removed. It was here to handle nicely the overflow; But, as a matter of fact, I don't care. If one wants to constraint them items with width and, gives a too long title; the label will just fall to its default size and will not impact the menu item